pyproject.toml
[project]
name = "scModelForge"
version = "0.1.0"
description = "A pretraining toolkit for single-cell foundation models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Ehsan" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anndata>=0.10",
    "scanpy>=1.9",
    "numpy>=1.23",
    "scipy>=1.9",
    "pandas>=1.5",
    "torch>=2.0",
    "pytorch-lightning>=2.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "rich>=13.0",
    "click>=8.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
census = ["cellxgene-census>=1.0"]
eval = ["scib>=1.1", "pertpy>=0.6"]
peft = ["peft>=0.6"]
wandb = ["wandb>=0.15"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "peft>=0.6",
    "ruff>=0.4",
    "mypy>=1.5",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=7.0",
    "myst-parser>=2.0",
    "sphinx-autodoc2>=0.5",
    "sphinx-book-theme>=1.0",
    "sphinx-copybutton>=0.5",
    "sphinx-design>=0.5",
]
all = ["scModelForge[census,eval,peft,wandb,dev,docs]"]

[project.scripts]
scmodelforge = "scmodelforge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scmodelforge"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["scmodelforge"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
]
